Feat(#100): 가상 스크롤 기능 구현 및 파일 구조 변경 - VirtualRender 가상 스크롤 기능 구현 - 폴더 구조 변경 및 코드 분리 - ChatPage => ChatRoomClient 파일 네이밍 변경
