--- src/api/chat/chat.clients.ts
@@ -1,27 +1,0 @@
-import { ChatType } from '@/types/chat/chat.type'
-import { ApiResponse } from '@/types/common'
-
-import { axios } from '../common'
-
-export const fetchChat = async (
-  params: {
-    page?: number
-    size?: number
-  },
-  chatRoomId: number,
-): Promise<ApiResponse<ChatType>> => {
-  const { page = 0, size = 30 } = params
-
-  const { data } = await axios.get<ApiResponse<ChatType>>(
-    `/v1/chats/${chatRoomId}/messages`,
-    {
-      params: {
-        page,
-        size,
-        sort: 'createdAt,desc',
-      },
-    },
-  )
-
-  return data
-}
