+++ src/app/chat/@chatRoom/[roomId]/page.tsx
@@ -1,5 +1,5 @@
-import ChatPage from '@/components/chat/ChatRoom'
+import ChatRoomClient from '@/features/chat/components/ChatRoomClient'
 
 export default function ChatRoom() {
-  return <ChatPage />
+  return <ChatRoomClient />
 }
