+++ src/app/chat/default.tsx
@@ -1,4 +1,4 @@
-import ChatListContainer from '@/components/chat/ChatList'
+import ChatListContainer from '@/features/chat/components/ChatListContainer/ChatListContainer'
 
 export default function DefaultChat() {
   return <ChatListContainer />
