+++ src/features/chat/api/chatroom.clients.ts
@@ -1,0 +1,29 @@
+import { axios } from '@/api/common'
+import { ChatRoomType } from '@/features/chat/types/chatRoom.types'
+import { ApiResponse } from '@/types/common'
+
+export const fetchChatRooms = async (params: {
+  page?: number
+  size?: number
+  keyword?: string
+}): Promise<ApiResponse<ChatRoomType>> => {
+  const { page = 0, size = 50, keyword } = params
+
+  const { data } = await axios.get<ApiResponse<ChatRoomType>>('/v1/chats', {
+    params: {
+      page,
+      size,
+      sort: 'createdAt,desc',
+      ...(keyword && { keyword }),
+    },
+  })
+
+  return data
+}
+export const leaveChatroom = async ({
+  chatParticipantId,
+}: {
+  chatParticipantId: number
+}): Promise<void> => {
+  await axios.patch(`/v1/chats/${chatParticipantId}/out`, {})
+}
